+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG
+## 0.45
+* Changed total work hours to client and internal work types (removed trust time from it)
+* Made a visual space to separate the work types from break and trust
+
 ## 0.44
 * Fixed empty type after using continue button.
 * Adding a "trust" type